Added pagination for search and fixed bugs
Fixed a bug in the FilmShortInfo class when the movie lacks either country or genre.
Now the FilmInfo method only returns existing information.

Pagination is now available for the search fragment.
For this, the FilmSearchPaginationFragment class was created.
For pagination on a fragment of top films, the TopFilmPaginationFragment class is used.